% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_prob_rf.R
\name{class_prob_rf}
\alias{class_prob_rf}
\title{Get Class Prediction Probabilities}
\usage{
class_prob_rf(rf_obj, newdata, breaks, labels, ordered = FALSE,
  type = c("probs", "votes"))
}
\arguments{
\item{rf_obj}{a randomForest object of type regression}

\item{newdata}{data frame used to make predictions. Can be same as original
data.}

\item{breaks}{a numeric vector of values used to turn predictions into classes}

\item{labels}{a character vector of labels for the classes}

\item{ordered}{logical indicating if categories are ordered (in order specified
in labels)}

\item{type}{a character vector indicating whether total votes or probabilites
should be returned}
}
\value{
returns a dataframe of probabilites or number of votes for each class
}
\description{
This function takes a regression randomForest and returns class probabilities
for a provided classifcation.  These probabilities are based on a full, new dataset.
If using the dataset used to build the random forest these are not out of
bag estimates.  The predicted classes that are output are directly from the random forest object,
thus they are predictions resulting from the out of bag esimates.
}

