% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condAccuracy.R
\name{condAccuracy}
\alias{condAccuracy}
\title{Plot conditional accuracy}
\usage{
condAccuracy(pred_prob1, pred_prob2, xlab = "x", ...)
}
\arguments{
\item{...}{parameters to be passed to condprob}

\item{rf}{A random forest object}
}
\description{
This function plots the total accuracy as a function of maximum probability.
We determine the maximum probability for each lake for a given random forest
object and then plot the total accuracy for lakes at each value of x or greater
}
\examples{
condAccuracy(all_rf_ts_prob,gis_rf_ts_prob,xImpair=0,R=1,xlab="Xc >= Prediction Probability")
}

