% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crossval_rf.R
\name{crossval_rf}
\alias{crossval_rf}
\title{Cross Validate Random Forests to get Categorical Accuracy}
\usage{
crossval_rf(y, x, breaks, cat = NULL, split, n, ntree)
}
\arguments{
\item{y}{response a vector}

\item{x}{predictors a data.frame}

\item{breaks}{numeric vector of cut points}

\item{cat}{categories of response}

\item{split}{proportion to include in training datset.  test set to inverse.}

\item{n}{number of iterations}
}
\description{
Funtion to cross-validate categorical prediction accuracy from
resultant chl a.  A Random forest of forests...  Used to get categorical
accuracy.  Also returns a cross-validated accuracy of % var explained.
}
\examples{
data(LakeTrophicModelling)
predictors_all <- predictors_all[predictors_all!="DATE_COL"]
all_cf_dat <- data.frame(ltmData[predictors_all],LogCHLA=log10(ltmData$CHLA))
all_cf_dat <- all_cf_dat[complete.cases(all_cf_dat),]
ts_brks <- c(min(all_cf_dat$LogCHLA),log10(2),log10(7),log10(30),max(all_cf_dat$LogCHLA))
ts_cats <- c("oligo","meso","eu","hyper")
x<-crossval_rf(all_cf_dat$LogCHLA,all_cf_dat[,names(all_cf_dat)!="LogCHLA"],
            ts_brks,ts_cats,0.8,10,10)
}

