% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{density plot for lake trophic modeling paper}
\usage{
density_plot(df, xvar, ...)
}
\description{
Produces a density plot via ggplot
}
\examples{
density_plot(gis_rf_ts_prob,xvar="max",x="Predicted Probability")
}
