% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dfToImage}
\alias{dfToImage}
\title{dfToImage}
\usage{
dfToImage(df, file, style = NULL, width = 1024, include.rownames = FALSE,
  keepHTML = FALSE, ...)
}
\arguments{
\item{df}{the data.frame to convert}

\item{file}{the output image file to save to include file type in the
extension.  Have tested , 'png', and 'jpg'}

\item{style}{the path/name for the style sheet.}

\item{width}{width of output image.  defaults to 1024.}

\item{include.rownames}{logical to print or not the row names.}

\item{keepHTML}{Logical if html file should be saved (best for troubleshooting
the css styling)}

\item{...}{parameters passed through to \code{\link{xtable}}.  For instance,
           allows use of \code{caption} option.}

\item{format.args}{list of formatC paramaters to control formating of text}
}
\description{
This function will take a data.frame, convert it to a table, style it with
CSS (optional) and save as an image file.
}
\examples{
dfToImage(head(mtcars),'test.jpg')
}

