% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecdf_ks_ci.R
\name{ecdf_ks_ci}
\alias{ecdf_ks_ci}
\title{Kolmogrov-Smirnoff CIs for a ecdf}
\usage{
ecdf_ks_ci(x)
}
\description{
Code from sfsmisc::ecdf.KSci. Ammended here to simply output the xy for both upper and
lower for use as input to ggplot
}
\examples{
ecdf_ks_ci(all_pred_prob)
}

