% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCyanoAbund}
\alias{getCyanoAbund}
\title{Pulls down cyano abundance from NLA website}
\usage{
getCyanoAbund()
}
\value{
a data.frame with two fields: SITE_ID=the NLA site id for the lake
         and cyanoCellsPerML=for each SITE_ID the sum of abundance (cells/ml)
         of all phytoplankton for Division='Cyanophyta'
}
\description{
This function downloads the NLA2007 phytoplankton data and calculates the
abundance (cells/ml) of cyanobacteria by lake
}

