% Generated by roxygen2 (4.0.2): do not edit by hand
\name{importancePlot}
\alias{importancePlot}
\title{Random Forest Variable Importance Plotting}
\usage{
importancePlot(rf, sumtbl, data_def = NULL, type = c("acc", "gini"), ...)
}
\arguments{
\item{rf}{a randomForest object}

\item{type}{the type of importance to plot "acc" for Mean Decrease Accuracy, and
"gini for Mean Decrease Gini.}

\item{...}{any items that can be passed to geom_point.  Mostly used to control size
           and color of points.}
}
\description{
Random Forest Variable Importance Plotting
}
\examples{
require(wesanderson)
data(iris)
iris_rf<-randomForest(iris[,1:4],iris[,5],importance=T,proximity=T)
importancePlot(iris_rf,'acc',size=10,aes(colour=wes.palette(4, 'GrandBudapest')))
importancePlot(iris_rf,'gini',size=10,aes(colour=wes.palette(4, 'GrandBudapest')))
}

