% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iterVarSelRF.R
\name{iterVarSelRF}
\alias{iterVarSelRF}
\title{iterVarSelRF}
\usage{
iterVarSelRF(indVar, depVar, numRun = 10, numCore = 1,
  outStr = tempfile(tmpdir = getwd()), ntree = 10, ntreeIterat = 5,
  vars.drop.frac = NULL, vars.drop.num = 1, time = FALSE)
}
\arguments{
\item{indVar}{independent variables}

\item{depVar}{dependent variable}

\item{numRun}{number of iterations.  default is 10}

\item{numCore}{number of cores to split runs across. defualt is 1}

\item{outStr}{string to append to output files. Defaults to a tmp file}

\item{time}{logical true will output overall time for the run.}
}
\description{
This function iterates the `varselRF` function to capture variability in the
possibly selected variables.  This process forms the basis of the analysis
conducted in Hollister, Kreakie, and Milsted (2014).  We us the  proportion
of iterations that a given variable is selected as a variable selection
metric. A parallel implementation is possible using `snowfall`.  User only need
specify the number of cores to use.  Function relies on knitr,snowfall,parallel,
and varSelRF
}
\examples{
data(LakeTrophicModelling)
x1<-iterVarSelRF(ltmData[predictors_gis],ltmData$TS_CHLA_4,10,ntree=10,ntreeIterat=5,
vars.drop.frac=NULL,vars.drop.num=1,time=TRUE)
x2<-iterVarSelRF(ltmData[predictors_gis],ltmData$TS_CHLA_4,10,4,ntree=10,ntreeIterat=5,
vars.drop.frac=NULL,vars.drop.num=1,time=TRUE)
}

