% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_prob_map.R
\name{make_ts_map}
\alias{make_ts_map}
\title{helper function to generate each map}
\usage{
make_ts_map(states, lake_prob, title, low, high)
}
\arguments{
\item{states}{the states}

\item{lake_prob}{merged lake and prob data}

\item{title}{title of map}

\item{low}{lower color}

\item{high}{higher color}
}
\description{
helper function to generate each map
}
