% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlaMap.R
\name{nlaMap}
\alias{nlaMap}
\title{Create map of NLA samples}
\usage{
nlaMap(states, lakes, cats = F, myColor = 1:3, catColor = NULL)
}
\arguments{
\item{states}{data for states, as data.frame}

\item{lakes}{point locations for lake samples as data.frame}

\item{cats}{plot ts categories?}

\item{myColor}{vector of length 3 with colors for fill, lines, and points in that order}
}
\description{
Function to generate Map 1 in hkm2014ESA poster
}
\examples{
state<-map_data('state')
lakes_alb<-data.frame(ltmData[["AlbersX"]],ltmData[["AlbersY"]])
p4s<-"+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
ll<-"+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs"
lakes_alb_sp<-SpatialPoints(coordinates(lakes_alb),proj4string=CRS(p4s))
lakes_dd<-spTransform(lakes_alb_sp,CRS=CRS(ll))
lakes_dd<-data.frame(coordinates(lakes_dd))
names(lakes_dd)<-c("long","lat")
nlaMap(state,lakes_dd)
}

