% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_plot.R
\name{partial_plot}
\alias{partial_plot}
\title{Create parital dependence plot}
\usage{
partial_plot(pd, xl, yl, ...)
}
\arguments{
\item{pd}{partial dependence from randomForest::partialPlot()}
}
\description{
This function takes the output of randomForest::paritalPlot() and
generates a plot via ggplot.
}
