% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runRandomForest}
\alias{runRandomForest}
\title{runRandomForest}
\usage{
runRandomForest(indVar, depVar, ntree = 100, importance = TRUE,
  proximity = TRUE)
}
\arguments{
\item{indVar}{independent variables}

\item{depVar}{dependent variable}

\item{ntree}{to be passed to randomForest.  Had trouble getting ... to work with the snowfall wrapper
function.  This was a quick hack.}

\item{importance}{to be passed to randomForest.  Had trouble getting ... to work with the snowfall wrapper
function.  This was a quick hack.}

\item{proximity}{to be passed to randomForest.  Had trouble getting ... to work with the snowfall wrapper
                 function.  This was a quick hack.}
}
\description{
This function takes the output of iterVarSelRF and runs a random forest.
}
\examples{
data(hkm2014Data)
x2<-iterVarSelRF(hkm2014Data[predictors_gis],TS_CHLA_4,10,4,ntree=10,ntreeIterat=5,
vars.drop.frac=NULL,vars.drop.num=1,time=FALSE)
x2_vars<-unique(unlist(x2))
gis_ts4_rf<-runRandomForest(hkm2014Data[x2_vars],TS_CHLA_4, ntree=10,importance=TRUE,proximity=TRUE)
}

