% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatterplot for hkm2014esa poster}
\usage{
scatterPlot(df, xvar, yvar, pt_col, ln_col, cl_col, ...)
}
\description{
Produces a scatterplot via ggplot
}
\examples{
devtools::install_github("wesanderson","jhollist")
library('wesanderson')
data(hkm2014Data)
mydf<-data.frame(chla=hkm2014Data[["CHLA"]],biovp1=hkm2014Data[["sumBioV"]]+1)
zissou2<-wes.palette(5,"Zissou2")
scatterPlot(mydf,xvar="chla",yvar="biovp1",zissou2[3],zissou[2],zissou[1],
            title="Chlorophyll and Cyanobacterai Relationship",
            x=expression(paste('Log10(Chl ', italic("a"),')')),
            y="Log10(Cyanobaterial Biovolumes + 1)")
}

