% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sumTable}
\alias{sumTable}
\title{sumTable}
\usage{
sumTable(tsList)
}
\arguments{
\item{tsList}{list of just the selected variable (i.e. truncate list by 5 if time=T in
               iterVarSelRF)}
}
\description{
Function to summarize the output of iterVarSelRF into a percentage of the time a given
variable was selected by varSelRF.  Outputs md via kable
}

