% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/varImportance.R
\name{varImportance}
\alias{varImportance}
\title{Variable Importance Summary}
\usage{
varImportance(rf_imp, label = NULL)
}
\arguments{
\item{rf_imp}{a Random Forest Importance Object}

\item{label}{A label for the model}
}
\description{
Variable Importance Summary
}

