% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsel_regression_rf.R
\name{varsel_regression_rf}
\alias{varsel_regression_rf}
\title{Variable Selection with Regression RF}
\usage{
varsel_regression_rf(y, x, prog = F, ...)
}
\arguments{
\item{y}{response a vector}

\item{x}{predictors a data.frame}

\item{...}{options to pass to randomForest}
}
\description{
Similar approach to varSelRF, but for regresssion.  Use full model to rank
variables based on either %IncMSE or Gini.  To use %IncMSE set 
imporatance = TRUE.  It then steps through that sorted variable list with 
most important first and runs RF, store variables and %MSE of model.
Note: this sorts on Gini, by default.  importance = TRUE ensures %IncMSE.
}
\examples{
data(LakeTrophicModelling)
predictors_all <- predictors_all[predictors_all!="DATE_COL"]
all_dat <- data.frame(ltmData[predictors_all],LogCHLA=log10(ltmData$CHLA))
all_dat <- all_dat[complete.cases(all_dat),]
x<-varsel_regression_rf(all_dat$LogCHLA,all_dat[,names(all_dat)!="LogCHLA"],
                        ntree=100,prog=T, importance = TRUE)
}
