% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotCdf}
\alias{plotCdf}
\title{plotCDF Function to create CDFs of chla, ts, and cyano}
\usage{
plotCdf(catVar, contVar, cdf_colors = 1:length(catVar), ...)
}
\arguments{
\item{catVar}{the categorical Variable such as-'bvCat', 'TS_CHLA_2', 'TS_CHLA_3', 'TS_CHLA_4'.
Of same length as contVar}

\item{contVar}{the continuous Variable such as-'sumBioV','CHLA', 'NTL', 'PTL', etc.  Of same
length as contVar}

\item{cdf_colors}{colors for CDF plots and median line -should equal # of categories in catVar + 1}

\item{...}{pass additional parameters to labs (e.g. set titles, axis labels, etc.)}
}
\description{
This functions produces CDF plots from the hkm2014Data.  The implementation may be used more
widely though to create cumulative distribution functions for multiple categories
}
\examples{
devtools::install_github('wesanderson','karthik')
library(wesanderson)
data(hkm2014Data)
plotCdf(hkm2014Data$bvCat,hkm2014Data$CHLA,cdf_colors=wes.palette(3,'Royal1'),y='Percent',
        x=expression(paste('Log10(Chl ', italic("a"),')')),title='CDF for category bvCat',
        color="Biovolume\\nCategories")
plotCdf(hkm2014Data$TS_CHLA_3,hkm2014Data$sumBioV+1,cdf_colors=wes.palette(3,'FantasticFox'),
        y='Percent',x='Log10(Cyanobacterial Biovolume)',
        title=expression(paste('CDF for Chlorophyll ', italic("a"),' Trophic States (3 Categories)')),
        color="Trophic State\\nCategories")
}

