% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob_cdf.R
\name{prob_cdf}
\alias{prob_cdf}
\title{Function to create CDFs of probability for each trophic state prediction}
\usage{
prob_cdf(probs1, probs2, ...)
}
\arguments{
\item{probs1}{prediction probabilities for one of the models}

\item{probs2}{prediction probabilities for the other model}

\item{...}{pass additional parameters to labs (e.g. set titles, axis labels, etc.)}
}
\description{
This functions produces CDF plots from predicted trophic state probabilities
}
\examples{
devtools::install_github('wesanderson','karthik')
library(wesanderson)
data(LakeTrophicModelling)
prob_cdf(all_rf_ts_prob, gis_rf_ts_prob,x="Prediction Probability",y="Proportion of Samples")
}

