% Generated by roxygen2 (4.0.1): do not edit by hand
\name{varImportance}
\alias{varImportance}
\title{Variable Importance Summary}
\usage{
varImportance(rf, label = NULL)
}
\arguments{
\item{rf}{Random Forest Object}

\item{label}{A label for the model}
}
\description{
Variable Importance Summary
}

